Migrate to storing email data in minio. #1

+++ app/Console/Commands/ProcessDeletedAccounts.php
@@ -2,12 +2,16 @@
 
 namespace App\Console\Commands;
 
+use App\Helpers\S3Helper;
 use App\Models\Account;
+use App\Models\Folder;
 use App\SearchProvider\MeilisearchProvider;
 use App\Services\AccountService;
+use Aws\S3\S3Client;
 use Illuminate\Console\Command;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\Storage;
 use Log;
 use LogicException;
 
@@ -41,14 +45,46 @@ class ProcessDeletedAccounts extends Command
         foreach ($deletedAccounts as $account) {
              Log::info('Processing account id ' . $account->id);
 
-             if ($this->deleteDatabaseEntries($account)) {
-                 Log::info('Deleted database entries for account id ' . $account->id);
+             if ($this->deleteS3Objects($account->user_id, $account->folders)) {
+                 Log::info('Deleted objects from S3 for account id ' . $account->id);
+
+                 if ($this->deleteDatabaseEntries($account)) {
+                     Log::info('Deleted database entries for account id ' . $account->id);
+                 } else {
+                     Log::error('Failed to delete database entries for account id ' . $account->id);
+                 }
 
                  $this->deletedSearchIndexEntries($account->folders, $account->user_id);
+             } else {
+                 Log::error('Failed to delete objects from S3 for account id ' . $account->id);
              }
         }
 
         return self::SUCCESS;
+    }
+
+    /**
+     * @param int $userId
+     * @param Collection $folders
+     * @return bool
+     */
+    private function deleteS3Objects(int $userId, Collection $folders): bool
+    {
+        /**
+         * @var $s3Client S3Client
+         */
+        $s3Client = Storage::disk('s3')->getClient();
+
+        /**
+         * @var $folder Folder
+         */
+        foreach ($folders as $folder) {
+            if (!S3Helper::deletePath($s3Client, $userId, $folder->id)) {
+                return false;
+            }
+        }
+
+        return true;
     }
 
     /**
